cmd: bootefi: rework do_bootefi()
In this patch, do_bootefi() will be reworked, without any functional
change, as it is a bit sloppy after Heinrich's "efi_loader: rework
loading and starting of images."

Under the current implementation, efi_selftest is quite irregular
in terms of loading an image. So it should be isolated from other
parts for readability and maintainability.

Signed-off-by: AKASHI Takahiro <takahiro.akashi@linaro.org>