env: extend interfaces allowing for env contexts
The notion of "env context" (or simply "context") will provide users
of env interfaces to manage their own domain of variables, which is
separated from other contexts and can be loaded/stored with dedicated
backing storage device.

With this patch, almost of all env interfaces are extended to accept
additional argument of a pointer to env context to identify targeted
domain after Wolfgang's requirement.

Optionally, a context may have no backing storage, in this case,
all the entries associated with this context will be lost after reboot
and env_save() returns -ENODEV.

In addition, "env" command is also modified to allow for specifying
a context with "-C <context name>" option.
Please note "-e" in "env set" and "env print" is still maintained
not only for backward compatibility but also for usefullness
at non-U-Boot-environment-based UEFI variables implementation,
in particular, secure storage solution in the future.

Signed-off-by: AKASHI Takahiro <takahiro.akashi@linaro.org>